requirements/development.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements/development.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-allauth
babel==2.17.0
    # via -r requirements/base.in
bandit==1.8.6
    # via -r requirements/development.in
beautifulsoup4==4.13.4
    # via -r requirements/base.in
black==25.1.0
    # via -r requirements/development.in
build==1.2.2.post1
    # via pip-tools
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   pip-tools
djade==1.4.0
    # via -r requirements/development.in
django==5.2.4
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   django-appconf
    #   django-ckeditor
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-js-asset
    #   django-otp
    #   django-polymorphic
    #   django-taggit
django-allauth==65.10.0
    # via -r requirements/base.in
django-any-urlfield==2.7.post1
    # via -r requirements/base.in
django-appconf==1.1.0
    # via django-imagekit
django-ckeditor==6.7.3
    # via -r requirements/base.in
django-debug-toolbar==5.2.0
    # via -r requirements/development.in
django-extensions==4.1
    # via -r requirements/development.in
django-filter==25.1
    # via -r requirements/base.in
django-imagekit==5.0.0
    # via -r requirements/base.in
django-js-asset==3.1.2
    # via
    #   django-ckeditor
    #   django-mptt
django-mptt==0.17.0
    # via django-polymorphic-tree
django-otp==1.6.1
    # via -r requirements/base.in
django-polymorphic==4.1.0
    # via django-polymorphic-tree
django-polymorphic-tree==2.1
    # via -r requirements/base.in
django-sortedm2m==4.0.0
    # via -r requirements/base.in
django-tag-parser==3.2
    # via django-polymorphic-tree
django-taggit==6.1.0
    # via django-taggit-autosuggest
django-taggit-autosuggest==0.4.2
    # via -r requirements/base.in
django-upgrade==1.25.0
    # via -r requirements/development.in
et-xmlfile==2.0.0
    # via openpyxl
factory-boy==3.3.3
    # via -r requirements/development.in
faker==37.4.2
    # via factory-boy
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.1.0
    # via black
openpyxl==3.1.5
    # via -r requirements/base.in
packaging==25.0
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pbr==6.1.1
    # via stevedore
pilkit==3.0
    # via django-imagekit
pillow==11.3.0
    # via pilkit
pip-tools==7.4.1
    # via -r requirements/development.in
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
psycopg[binary,pool]==3.2.9
    # via -r requirements/base.in
psycopg-binary==3.2.9
    # via psycopg
psycopg-pool==3.2.6
    # via psycopg
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via
    #   -r requirements/development.in
    #   pytest-django
    #   pytest-random-order
pytest-django==4.11.1
    # via -r requirements/development.in
pytest-random-order==1.2.0
    # via -r requirements/development.in
pyupgrade==3.20.0
    # via -r requirements/development.in
pyyaml==6.0.2
    # via bandit
requests==2.32.4
    # via -r requirements/base.in
rich==14.0.0
    # via bandit
rules==3.5
    # via -r requirements/base.in
soupsieve==2.7
    # via beautifulsoup4
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
stevedore==5.4.1
    # via bandit
structlog==25.4.0
    # via -r requirements/base.in
tokenize-rt==6.2.0
    # via
    #   django-upgrade
    #   pyupgrade
typeguard==4.4.4
    # via -r requirements/base.in
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   psycopg-pool
    #   typeguard
tzdata==2025.2
    # via faker
urllib3==2.5.0
    # via requests
wheel==0.45.1
    # via pip-tools
xlrd==2.0.2
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
